# openapi/public.yaml
openapi: 3.0.1
info:
  title: CGN Onboarding Portal Public API
  description: Carta Giovani Nazionale Onboarding Portal - Public API
  version: 0.1.0
servers:
  - url: https://api.cgn.pagopa.it/public/api
tags:
  - name: help
    description: Help
paths:
  /session:
    post:
      summary: creates a user session
      operationId: createJwtSessionToken
      description: >
        This endpoint accepts two alternative forms of data:
        - OneIdentityData to login from One Identity (code + state)
        - ActiveDirectoryData to login from Active Directory (pagopa_token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/OneIdentityData'
                - $ref: '#/components/schemas/ActiveDirectoryData'
              discriminator:
                propertyName: requestType
                mapping:
                  oi: '#/components/schemas/OneIdentityData'
                  ad: '#/components/schemas/ActiveDirectoryData'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Parameters are not valid
  /help:
    post:
      tags:
        - help
      summary: Send an help request
      operationId: sendHelpRequest
      requestBody:
        description: Help request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpRequest'
        required: true
      responses:
        "204":
          description: Request Received
          content: { }
      x-codegen-request-body-name: helpRequest
components:
  schemas:
    OneIdentityData:
      type: object
      required:
        - code
        - state
        - nonce
        - requestType
      properties:
        code:
          type: string
          description: code received from One Identity
        state:
          type: string
          description: state received from One Identity
        nonce:
          type: string
          description: random field
        requestType:
          type: string
          description: discriminator field
    
    ActiveDirectoryData:
      type: object
      required:
        - token
        - requestType
      properties:
        token:
          type: string
          description: Authentication token received from Active Directory
        requestType:
          type: string
          description: discriminator field

    HelpRequest:
      required:
        - category
        - emailAddress
        - legalName
        - message
        - recaptchaToken
        - referentFirstName
        - referentLastName
      type: object
      properties:
        legalName:
          type: string
        referentFirstName:
          type: string
        referentLastName:
          type: string
        emailAddress:
          type: string
          format: email
        category:
          type: string
          enum:
            - Access
            - DataFilling
            - Discounts
            - Documents
            - TechnicalProblem
            - CgnOwnerReporting
            - Suggestions
            - Other
        topic:
          type: string
        message:
          maxLength: 200
          type: string
        recaptchaToken:
          type: string
  responses:
    InvalidRequest:
      description: Bad request
      content: { }
    Forbidden:
      description: Forbidden
      content: { }
    NotFound:
      description: Not found
      content: { }
x-original-swagger-version: "2.0"